requirements.txt
demucs==4.0.1
fastapi==0.116.1
requests==2.31.0
firebase-admin==6.9.0
uvicorn==0.35.0
yt-dlp==2025.7.13.233101.dev0
python-dotenv==1.0.0
schedule==1.2.0